gui is af, alleen 3mxl control toevogen aan het jassen algoritme
