--- packages/garderobot/garderobot/errors.py
@@ -1,6 +1,0 @@
-class WardrobeError(Exception):
-    def __init__(self, value: str):
-        self.value = value
-
-    def __str__(self):
-        return repr(self.value)
